+++ frontend/src/components/BottomTabBar/index.tsx
@@ -3,8 +3,8 @@ import { TabBarContainer, TapBarItem } from './style';
 import { AiOutlineHome } from '@react-icons/all-files/ai/AiOutlineHome';
 import { AiOutlineGift } from '@react-icons/all-files/ai/AiOutlineGift';
 import { AiOutlineUser } from '@react-icons/all-files/ai/AiOutlineUser';
-import { RouterPath } from '../../../../types/utils';
-import ROUTER_PATH from '../../../../constants/routerPath';
+import { RouterPath } from '../../types/utils';
+import ROUTER_PATH from '../../constants/routerPath';
 import { FaStamp } from '@react-icons/all-files/fa/FaStamp';
 import { FaListUl } from '@react-icons/all-files/fa/FaListUl';
 import { FaMedal } from '@react-icons/all-files/fa/FaMedal';
